requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12
pydantic==2.9.2
PyPDF2==3.0.1
python-docx==1.1.0
transformers==4.46.0
torch==2.5.1
sentencepiece==0.2.0
yake==0.4.8
spacy==3.8.2
pdfplumber==0.11.4
aiofiles==24.1.0
